Improve docker files for immutable prod and live dev
Rewrote the docker related files to produce small, reproducible images
while still allowing for the previous development type of image with a
supplemental docker-compose file. While this is suitable for use on the
docker hub it still takes some effort due to the settings.ini which
should have its values be overridable via environment variable.